BREAKING: Use Vite as our builder
