+++ .storybook/main.ts
@@ -1,4 +1,4 @@
-import type { StorybookConfig } from "@storybook/react-webpack5";
+import type { StorybookConfig } from "@storybook/react-vite";
 
 const config: StorybookConfig = {
   stories: ["../src/**/*.story.@(jsx|tsx)"],
@@ -16,16 +16,24 @@ const config: StorybookConfig = {
     },
     "@storybook/addon-a11y",
     "@storybook/addon-actions",
-    "@storybook/addon-webpack5-compiler-babel",
-    "@chromatic-com/storybook",
   ],
 
-  framework: "@storybook/react-webpack5",
+  framework: {
+    name: "@storybook/react-vite",
+    options: {},
+  },
+
+  async viteFinal(config) {
+    // Merge with the main Vite config if needed
+    // Storybook will use its own Vite config, but we can customize it here
+    return config;
+  },
 
   docs: {},
 
   typescript: {
-    reactDocgen: "react-docgen-typescript",
+    reactDocgen: false,
+    check: false,
   },
 };
