+++ package.json
@@ -13,9 +13,9 @@
     "start:cypress": "start-server-and-test start 9999 'cypress open'",
     "start:storybook": "storybook dev -p 9999",
     "clean:storybook": "rm -rf node_modules/.cache/storybook && rm -rf ./storybook-static",
-    "build": "rollup -c",
-    "build:watch": "rollup -c --watch",
-    "build:storybook": "storybook build --webpack-stats-json",
+    "build": "vite build",
+    "build:watch": "vite build --watch",
+    "build:storybook": "storybook build --stats-json",
     "warn:prepush": "echo \"Make sure you also run all the other CI steps before pushing by running 'pnpm ci'\"",
     "check": "pnpm warn:prepush && pnpm check:types && pnpm check:lint && pnpm check:format",
     "check:types": "tsc && cd cypress && tsc --noEmit",
@@ -61,17 +61,8 @@
   },
   "devDependencies": {
     "@apollo/client": "^3.11.8",
-    "@babel/cli": "^7.28.3",
-    "@babel/core": "^7.28.5",
-    "@babel/plugin-transform-runtime": "^7.28.5",
-    "@babel/preset-env": "^7.28.5",
-    "@babel/preset-react": "^7.28.5",
-    "@babel/preset-typescript": "^7.28.5",
-    "@chromatic-com/storybook": "^3",
     "@nulogy/eslint-config-nulogy": "^1.0.0",
     "@nulogy/icons": "^4.37.2",
-    "@rollup/plugin-babel": "^5.0.0",
-    "@rollup/plugin-node-resolve": "^7.1.3",
     "@semantic-release/changelog": "^6.0.2",
     "@semantic-release/commit-analyzer": "^9.0.2",
     "@semantic-release/git": "^10.0.1",
@@ -84,10 +75,9 @@
     "@storybook/addon-storysource": "8.6.14",
     "@storybook/addon-toolbars": "8.6.14",
     "@storybook/addon-viewport": "8.6.14",
-    "@storybook/addon-webpack5-compiler-babel": "^4.0.0",
     "@storybook/codemod": "8.6.14",
     "@storybook/react": "8.6.14",
-    "@storybook/react-webpack5": "8.6.14",
+    "@storybook/react-vite": "8.6.14",
     "@storybook/theming": "8.6.14",
     "@testing-library/react": "^12.1.5",
     "@types/css-mediaquery": "^0.1.4",
@@ -95,15 +85,10 @@
     "@types/node": "^24.0.0",
     "@types/react": "^17.0.39",
     "@types/react-dom": "^17.0.20",
-    "@types/react-router-dom": "^5.3.3",
+    "@types/react-router-dom": "5.3.0",
     "@typescript-eslint/eslint-plugin": "^4.0.0",
     "@typescript-eslint/parser": "^5.30.5",
-    "babel-eslint": "^10.0.3",
-    "babel-loader": "8.0.6",
-    "babel-plugin-inline-json-import": "^0.3.2",
-    "babel-plugin-require-context-hook": "1.0.0",
-    "babel-plugin-styled-components": "^2.1.4",
-    "babel-preset-react": "6.24.1",
+    "@vitejs/plugin-react": "^4.7.0",
     "chromatic": "^13.3.3",
     "concurrently": "^9.2.0",
     "css-mediaquery": "^0.1.2",
@@ -124,22 +109,16 @@
     "react-color": "^2.18.1",
     "react-dom": "17.0.2",
     "react-router-dom": "^5.3.3",
-    "rollup": "^2.7.3",
-    "rollup-plugin-babel": "^4.4.0",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-replace": "^2.2.0",
-    "rollup-plugin-typescript2": "^0.36.0",
     "semantic-release": "^24.2.5",
     "start-server-and-test": "^2.0.12",
     "storybook": "8.6.14",
     "styled-components": "^6.1.19",
     "svg-sprite-loader": "^6.0.11",
     "typescript": "^5.9.3",
-    "typescript-eslint": "^0.0.1-alpha.0",
-    "vitest": "^4.0.15"
+    "vite": "^5.0.0",
+    "vitest": "^2.0.0"
   },
   "dependencies": {
-    "@babel/runtime": "^7.9.6",
     "@emotion/is-prop-valid": "^1.3.1",
     "@nulogy/tokens": "^6.1.1",
     "@radix-ui/react-navigation-menu": "^1.1.4",
@@ -149,7 +128,6 @@
     "@styled-system/theme-get": "^5.1.2",
     "@types/react-window": "^1.8.8",
     "@types/styled-system": "5.1.25",
-    "core-js": "3",
     "date-fns": "2.23.0",
     "debounce": "^1.2.0",
     "deep-equal": "^2.2.1",
