+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "target": "ES2020",
+    "target": "ES2015",
     "declaration": true,
     "outDir": "./dist",
     "lib": ["dom", "dom.iterable", "esnext"],
